requirements.txt
aiohttp
aiofiles
asgiref==3.8.1
celery==5.4.0
channels
channels-redis
Django==5.1.5
django-autoslug
django-celery-beat
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-nested-routers
drf-spectacular==0.28.0
fastjsonschema==2.21.1
flower==2.0.1
freezegun==1.5.1
gunicorn
humanize==4.11.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
logging
loguru==0.7.3
nest-asyncio==1.6.0
oauthlib==3.2.2
platformdirs==4.3.6
prometheus_client==0.21.1
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
typing_extensions==4.12.2
websockets
websocket-client==1.8.0
pika==1.3.2
aio-pika==9.5.4
